+++ lib/moosyl.dart
@@ -1,5 +1,4 @@
 export 'src/pages/home.dart';
 export 'src/models/payment_method_model.dart';
-export 'src/models/payment_success.dart';
 export 'src/moosyl_flutter.dart';
 export 'l10n/moosyl_localization.dart';