+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Remove node.js Example
 - update the returned payment model desc
+- payment view return bool instead of model
 
 ## 2.0.1